+++ glom/test/test_basic.py
@@ -1,7 +1,12 @@
 
+import sys
+from xml.etree import cElementTree as ElementTree
+
 import pytest
 
-from glom import glom, SKIP, STOP, Path, Inspect, Coalesce, CoalesceError, Literal, Call, T, S, Spec
+from glom import glom, SKIP, STOP, Path, Inspect, Coalesce, CoalesceError, Literal, Call, T, S, Invoke, Spec, Ref
+from glom import Auto, Fill, Iter
+
 import glom.core as glom_core
 from glom.core import UP, ROOT, Let
 
@@ -77,7 +82,9 @@ def test_coalesce():
     assert expected.replace(',', '') in received.replace(',', '')  # normalize commas for py3.7+ repr change
 
     # check that defaulting works
-    assert glom(val, Coalesce('xxx', 'yyy', default='zzz')) == 'zzz'
+    spec = Coalesce('xxx', 'yyy', default='zzz')
+    assert glom(val, spec) == 'zzz'
+    assert repr(spec) == "Coalesce('xxx', 'yyy', default='zzz')"
 
     # check that default_factory works
     sentinel_list = []
@@ -92,6 +99,7 @@ def test_coalesce():
 
     # check that arbitrary exceptions can be ignored
     assert glom(val, Coalesce(lambda x: 1/0, 'a.b', skip_exc=ZeroDivisionError)) == 'c'
+
 
 
 def test_skip():
@@ -173,8 +181,10 @@ def test_abstract_iterable():
     class MyIterable(object):
         def __iter__(self):
             return iter([1, 2, 3])
+    mi = MyIterable()
+    assert list(mi) == [1, 2, 3]
 
-    assert isinstance(MyIterable(), glom_core._AbstractIterable)
+    assert isinstance(mi, glom_core._AbstractIterable)
 
 
 def test_call_and_target():
@@ -200,7 +210,74 @@ def test_call_and_target():
 
     with pytest.raises(TypeError, match='expected func to be a callable or T'):
         Call(func=object())
+
+    assert glom(lambda: 'hi', Call()) == 'hi'
     return
+
+
+def test_invoke():
+    args = []
+    def test(*a, **kw):
+        args.append(a)
+        args.append(kw)
+        return 'test'
+
+    assert glom('a', Invoke(len).specs(T)) == 1
+    data = {
+        'args': (1, 2),
+        'args2': (4, 5),
+        'kwargs': {'a': 'a'},
+        'c': 'C',
+    }
+    spec = Invoke(test).star(args='args'
+        ).constants(3, b='b').specs(c='c'
+        ).star(args='args2', kwargs='kwargs')
+    repr(spec)  # no exceptions
+    assert repr(Invoke(len).specs(T)) == 'Invoke(len).specs(T)'
+    assert (repr(Invoke.specfunc(next).constants(len).constants(1))
+            == 'Invoke.specfunc(next).constants(len).constants(1)')
+    assert glom(data, spec) == 'test'
+    assert args == [
+        (1, 2, 3, 4, 5),
+        {'a': 'a', 'b': 'b', 'c': 'C'}]
+    args = []
+    assert glom(test, Invoke.specfunc(T)) == 'test'
+    assert args == [(), {}]
+    repr_spec = Invoke.specfunc(T).star(args='args'
+        ).constants(3, b='b').specs(c='c'
+        ).star(args='args2', kwargs='kwargs')
+    assert repr(eval(repr(repr_spec), locals(), globals())) == repr(repr_spec)
+
+    with pytest.raises(TypeError, match='expected func to be a callable or Spec instance'):
+        Invoke(object())
+    with pytest.raises(TypeError, match='expected one or both of args/kwargs'):
+        Invoke(T).star()
+
+    # test interleaved pos args
+    def ret_args(*a, **kw):
+        return a, kw
+
+    spec = Invoke(ret_args).constants(1).specs({}).constants(3)
+    assert glom({}, spec) == ((1, {}, 3), {})
+    # .endswith because ret_arg's repr includes a memory location
+    assert repr(spec).endswith(').constants(1).specs({}).constants(3)')
+
+    # test overridden kwargs
+    should_stay_empty = []
+    spec = Invoke(ret_args).constants(a=1).specs(a=should_stay_empty.append).constants(a=3)
+    assert glom({}, spec) == ((), {'a': 3})
+    assert len(should_stay_empty) == 0
+    assert repr(spec).endswith(').constants(a=3)')
+
+    # bit of coverage
+    target = (lambda: 'hi', {})
+    spec = Invoke(T[0])
+    assert glom(target, spec) == 'hi'
+    # and a bit more
+    spec = spec.star(kwargs=T[1])
+    assert repr(spec) == 'Invoke(T[0]).star(kwargs=T[1])'
+    assert glom(target, spec) == 'hi'
+
 
 
 def test_spec_and_recursion():
@@ -326,10 +403,49 @@ def test_inspect():
 def test_let():
     data = {'a': 1, 'b': [{'c': 2}, {'c': 3}]}
     output = [{'a': 1, 'c': 2}, {'a': 1, 'c': 3}]
-    assert glom(data, Let(a='a').over(('b', [{'a': S['a'], 'c': 'c'}]))) == output
+    assert glom(data, (Let(a='a'), ('b', [{'a': S['a'], 'c': 'c'}]))) == output
     assert glom(data, ('b', [{'a': S[ROOT][Literal(T)]['a'], 'c': 'c'}])) == output
 
     with pytest.raises(TypeError):
         Let('posarg')
     with pytest.raises(TypeError):
         Let()
+
+    assert repr(Let(a=T.a.b)) == 'Let(a=T.a.b)'
+
+
+def test_ref():
+    assert glom([[[]]], Ref('item', [Ref('item')])) == [[[]]]
+    with pytest.raises(Exception):  # check that it recurses downards and dies on int iteration
+        glom([[[1]]], Ref('item', [Ref('item')]))
+    assert repr(Ref('item', (T[1], Ref('item')))) == "Ref('item', (T[1], Ref('item')))"
+
+    etree2dicts = Ref('ElementTree',
+        {"tag": "tag", "text": "text", "attrib": "attrib", "children": (iter, [Ref('ElementTree')])})
+    etree2tuples = Fill(Ref('ElementTree', (T.tag, Iter(Ref('ElementTree')).all())))
+    etree = ElementTree.fromstring('''
+    <html>
+      <head>
+        <title>the title</title>
+      </head>
+      <body id="the-body">
+        <p>A paragraph</p>
+      </body>
+    </html>''')
+    glom(etree, etree2dicts)
+    glom(etree, etree2tuples)
+
+
+_IS_PYPY = '__pypy__' in sys.builtin_module_names
+@pytest.mark.skipif(_IS_PYPY, reason='pypy othertype.__repr__ is never object.__repr__')
+def test_api_repr():
+    import glom
+
+    spec_types_wo_reprs = []
+    for k, v in glom.__dict__.items():
+        if not callable(getattr(v, 'glomit', None)):
+            continue
+        if v.__repr__ is object.__repr__:
+            spec_types_wo_reprs.append(k)  # pragma: no cover
+
+    assert set(spec_types_wo_reprs) == set([])